stumped
Turns out, adaboost by default uses decision trees of height 1 as their classifier. This will slow runtime, but it'll be much more accurate if you use decision tree stumps that are fitted as the classifiers. I cut the number of estimators from 200 to 100 to help runtime, and the stumps are height 5 with 2 samples minimum per leaf, which is pretty easy on runtime as well. more accuracy and runtime means larger height and smaller min samples per leaf